requirements.txt
pandas==1.5.3
numpy==1.24.3
matplotlib==3.7.1
tensorflow==2.13.0
streamlit==1.22.0
tensorflow==2.13.0
numpy==1.24.3
Pillow==9.5.0
